Report number of Git stashes
To avoid unnecessary parsing, we access the .git/ref/stash file
directly. Each line represents a separate stash, so a simple line count
(including zero) is all we need. This value is concatenated to the end
of the _GIT_STATUS line to simplify the parsing with one common
technique for all fields.

Note: This feature is active for both the Python and Haskell versions.